+++ src/main/java/fr/the_gacha_company/projet_r304_gacha/hero/Gender.java
@@ -1,4 +1,4 @@
-package fr.the_gacha_company.projet_r304_gacha;
+package fr.the_gacha_company.projet_r304_gacha.hero;
 
 public enum Gender {
