+++ src/main/java/fr/the_gacha_company/projet_r304_gacha/hero/heroes/Godric.java
@@ -1,9 +1,9 @@
-package fr.the_gacha_company.projet_r304_gacha.heroes;
+package fr.the_gacha_company.projet_r304_gacha.hero.heroes;
 
-import fr.the_gacha_company.projet_r304_gacha.Gender;
-import fr.the_gacha_company.projet_r304_gacha.Rarity;
-import fr.the_gacha_company.projet_r304_gacha.Stat;
-import fr.the_gacha_company.projet_r304_gacha.races.Human;
+import fr.the_gacha_company.projet_r304_gacha.hero.Gender;
+import fr.the_gacha_company.projet_r304_gacha.hero.Rarity;
+import fr.the_gacha_company.projet_r304_gacha.hero.Stat;
+import fr.the_gacha_company.projet_r304_gacha.hero.races.Human;
 
 public class Godric extends Human {
