+++ src/main/java/fr/the_gacha_company/projet_r304_gacha/hero/races/Human.java
@@ -1,9 +1,9 @@
-package fr.the_gacha_company.projet_r304_gacha.races;
+package fr.the_gacha_company.projet_r304_gacha.hero.races;
 
-import fr.the_gacha_company.projet_r304_gacha.Gender;
-import fr.the_gacha_company.projet_r304_gacha.Hero;
-import fr.the_gacha_company.projet_r304_gacha.Rarity;
-import fr.the_gacha_company.projet_r304_gacha.Stat;
+import fr.the_gacha_company.projet_r304_gacha.hero.Gender;
+import fr.the_gacha_company.projet_r304_gacha.hero.Hero;
+import fr.the_gacha_company.projet_r304_gacha.hero.Rarity;
+import fr.the_gacha_company.projet_r304_gacha.hero.Stat;
 
 public abstract class Human extends Hero {
