+++ src/main/java/fr/the_gacha_company/projet_r304_gacha/hero/roles/Warrior.java
@@ -1,6 +1,6 @@
-package fr.the_gacha_company.projet_r304_gacha.roles;
+package fr.the_gacha_company.projet_r304_gacha.hero.roles;
 
-import fr.the_gacha_company.projet_r304_gacha.Hero;
+import fr.the_gacha_company.projet_r304_gacha.hero.Hero;
 
 public interface Warrior {
